add guard to check for run_storage
Summary: Need to guard that this db is used for run_storage

Test Plan: bk

Reviewers: sashank, schrockn, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D3490

+++ python_modules/libraries/dagster-postgres/dagster_postgres/run_storage/alembic/versions/c63a27054f08_add_snapshots_to_run_storage.py
@@ -21,6 +21,9 @@ depends_on = None
 
 
 def upgrade():
+    if not has_table('runs'):
+        return
+
     if not has_table('snapshots'):
         op.create_table(
             'snapshots',
@@ -38,8 +41,11 @@ def upgrade():
 
 
 def downgrade():
-    if has_column('runs', 'snapshot_id'):
-        op.drop_column('runs', 'snapshot_id')
-
     if has_table('snapshots'):
         op.drop_table('snapshots')
+
+    if not has_table('runs'):
+        return
+
+    if has_column('runs', 'snapshot_id'):
+        op.drop_column('runs', 'snapshot_id')
